glibc: Remove site_config and glibc-initial
The only reason we appear to need glibc-initial at this time is to support
the site_config code. The site_config code compiles and therefore needs
at least some level of working C library to link against.

This isn't a good reason to keep the complexity of glibc-initial around
so remove it, and the site_config support which then breaks.

Performance benchmarks suggest the time spent just rerunning configure
for site_config just about equals any performance benefit for OE-Core
image builds excluding the time spent adding glibc-initial to the
dependency chain.

I'm not opposed to readding some other form of site_config support
but it needs to be rethought.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>