meta: use explicit uri for all sourceforge upstream version checks
Previously the check for some recipes relied on sourceforge redirecting from
downloads.sourceforge.net (SRC_URI is set to that) into the actual project page.

Sourceforge does this for interactive browsers, but not for wget.

With the check no longer mimicking a browser, and being truthful
about coming from wget we need to explicitly fetch
the project page in all cases, which is what this commit does.
(many recipes already set this explicitly and don't need to be tweaked)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>